+++ Comm2Tender/Logic/Models/Dto/Role.cs
@@ -7,7 +7,7 @@ namespace Comm2Tender.Logic.Models.Dto
 {
     public class Role
     {
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
         public string Name { get; set; }   
 
         public static implicit operator Role(Data.Role a)
